complete vite runtime asset closeout
