+++ src/fs/library-delivery.ts
@@ -1,11 +1,12 @@
 import type { AppMode } from '../state/url-mode.ts';
 import { resolveRuntimeAssetUrl } from '../runtime/asset-urls.ts';
+import { openSCADWasmUrl } from '../runner/openscad-asset-urls.ts';
 import { zipArchives, type ZipArchive } from './zip-archives.generated.ts';
 
 export const LIBRARY_DELIVERY_POLICY =
   'selected-prefetch bootstrap + editor eager mount + worker demand-load';
 
-const CORE_PREFETCH_SPECIFIERS = ['openscad.wasm', 'libraries/fonts.zip'];
+const CORE_PREFETCH_SPECIFIERS = [openSCADWasmUrl, 'libraries/fonts.zip'];
 
 export function getPrefetchedArchives(archives: ZipArchive[] = zipArchives): ZipArchive[] {
   return archives.filter((archive) => archive.prefetch === true);
