+++ src/runner/openscad-asset-urls.ts
@@ -1,0 +1,3 @@
+import openSCADWasmUrl from '../wasm/openscad.wasm?url';
+
+export { openSCADWasmUrl };
