+++ src/runner/openscad-runtime.ts
@@ -3,7 +3,7 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore — no type declarations for the WASM module
 import OpenSCAD from '../wasm/openscad.js';
-import { resolveRuntimeAssetUrl } from '../runtime/asset-urls.ts';
+import { openSCADWasmUrl } from './openscad-asset-urls.ts';
 
 export type RuntimeOptions = {
   print: (text: string) => void;
@@ -32,7 +32,7 @@ export async function createRuntime(opts: RuntimeOptions): Promise<OpenSCADRunti
     noInitialRun: true,
     print: opts.print,
     printErr: opts.printErr,
-    locateFile: (path: string) => resolveRuntimeAssetUrl(path),
+    locateFile: (path: string) => (path === 'openscad.wasm' ? openSCADWasmUrl : path),
   });
 
   return {
